+++ src/index.ts
@@ -14,3 +14,4 @@ export * from './probe';
 export * from './pvc';
 export * from './pv';
 export * from './handler';
+export * from './workload';
